requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.0
pillow==10.3.0
proglog==0.1.10
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
